Update where json files are saved
Rather than just dump all the json files we create into a single directory and ensure that navigating the directory is imposible, create a hierarchy of directories that starts with a directory of the region number, then a directory of the callsign prefix before the region number and then place all the json files for that region number and prefix into that directory. This ensures that the files are still easy to find but that we can navigate their directory hierachy easily. So for instance given the callsign K1ABC, the json file will now be in _build/callsigns/1/K/K1ABC.json