Turn off quoting in the python csv library
Recently I found an error in the FCC EN.dat file.  This record:

EN|3473987|||W2JJW|L|L01773257|West, James "|JAMES|"|WEST|||||602 PROVINCE LINE RD|ALLENTOWN|NJ|08501|||000|0022526404|I||||||

caused the csv parser to throw all kinds of errors due to the quotes inside the fields. To avoid these errors and because quoting is not used in the FCC dat files anyway, the solution is to simply tell the python csv library to turn quoting off.

For reference, the quotes in that line used to be the letter J. I'm not sure why they changed.